Module google_lifesciences2_beta::api

Structs

Accelerator
Carries information about an accelerator that can be attached to a VM.

Action
Specifies a single action that runs a Docker container.

CancelOperationRequest
The request message for Operations.CancelOperation.

CloudLifeSciences
Central instance to access all CloudLifeSciences related resource activities

Disk
Carries information about a disk that can be attached to a VM. See https://cloud.google.com/compute/docs/disks/performance for more information about disk type, size, and performance considerations. Specify either Volume or Disk, but not both.

Empty
A generic empty message that you can re-use to avoid defining duplicated empty messages in your APIs. A typical example is to use it as the request or the response type of an API method. For instance: service Foo { rpc Bar(google.protobuf.Empty) returns (google.protobuf.Empty); } The JSON representation for Empty is empty JSON object {}.

ExistingDisk
Configuration for an existing disk to be attached to the VM.

ListLocationsResponse
The response message for Locations.ListLocations.

ListOperationsResponse
The response message for Operations.ListOperations.

Location
A resource that represents Google Cloud Platform location.

Mount
Carries information about a particular disk mount inside a container.

NFSMount
Configuration for an NFSMount to be attached to the VM.

Network
VM networking options.

Operation
This resource represents a long-running operation that is the result of a network API call.

PersistentDisk
Configuration for a persistent disk to be attached to the VM. See https://cloud.google.com/compute/docs/disks/performance for more information about disk type, size, and performance considerations.

Pipeline
Specifies a series of actions to execute, expressed as Docker containers.

ProjectLocationGetCall
Gets information about a location.

ProjectLocationListCall
Lists information about the supported locations for this service.

ProjectLocationOperationCancelCall
Starts asynchronous cancellation on a long-running operation. The server makes a best effort to cancel the operation, but success is not guaranteed. Clients may use Operations.GetOperation or Operations.ListOperations to check whether the cancellation succeeded or the operation completed despite cancellation. Authorization requires the following Google IAM permission: * lifesciences.operations.cancel

ProjectLocationOperationGetCall
Gets the latest state of a long-running operation. Clients can use this method to poll the operation result at intervals as recommended by the API service. Authorization requires the following Google IAM permission: * lifesciences.operations.get

ProjectLocationOperationListCall
Lists operations that match the specified filter in the request. Authorization requires the following Google IAM permission: * lifesciences.operations.list

ProjectLocationPipelineRunCall
Runs a pipeline. The returned Operation's metadata field will contain a google.cloud.lifesciences.v2beta.Metadata object describing the status of the pipeline execution. The response field will contain a google.cloud.lifesciences.v2beta.RunPipelineResponse object if the pipeline completes successfully. Note: Before you can use this method, the Life Sciences Service Agent must have access to your project. This is done automatically when the Cloud Life Sciences API is first enabled, but if you delete this permission you must disable and re-enable the API to grant the Life Sciences Service Agent the required permissions. Authorization requires the following Google IAM permission: * lifesciences.workflows.run

ProjectMethods
A builder providing access to all methods supported on project resources. It is not used directly, but through the CloudLifeSciences hub.

Resources
The system resources for the pipeline run. At least one zone or region must be specified or the pipeline run will fail.

RunPipelineRequest
The arguments to the RunPipeline method. The requesting user must have the iam.serviceAccounts.actAs permission for the Cloud Life Sciences service account or the request will fail.

Secret
Holds encrypted information that is only decrypted and stored in RAM by the worker VM when running the pipeline.

ServiceAccount
Carries information about a Google Cloud service account.

Status
The Status type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by gRPC. Each Status message contains three pieces of data: error code, error message, and error details. You can find out more about this error model and how to work with it in the API Design Guide.

VirtualMachine
Carries information about a Compute Engine VM resource.

Volume
Carries information about storage that can be attached to a VM. Specify either Volume or Disk, but not both.

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.