Struct google_lifesciences2_beta::api::Action

pub struct Action {
    pub always_run: Option<bool>,
    pub block_external_network: Option<bool>,
    pub commands: Option<Vec<String>>,
    pub container_name: Option<String>,
    pub credentials: Option<Secret>,
    pub disable_image_prefetch: Option<bool>,
    pub disable_standard_error_capture: Option<bool>,
    pub enable_fuse: Option<bool>,
    pub entrypoint: Option<String>,
    pub environment: Option<HashMap<String, String>>,
    pub ignore_exit_status: Option<bool>,
    pub image_uri: Option<String>,
    pub labels: Option<HashMap<String, String>>,
    pub mounts: Option<Vec<Mount>>,
    pub pid_namespace: Option<String>,
    pub port_mappings: Option<HashMap<String, i32>>,
    pub publish_exposed_ports: Option<bool>,
    pub run_in_background: Option<bool>,
    pub timeout: Option<String>,
}

Specifies a single action that runs a Docker container.

This type is not used in any activity, and only used as part of another schema.

Fields
always_run: Option<bool>

By default, after an action fails, no further actions are run. This flag indicates that this action must be run even if the pipeline has already failed. This is useful for actions that copy output files off of the VM or for debugging. Note that no actions will be run if image prefetching fails.

block_external_network: Option<bool>

Prevents the container from accessing the external network.

commands: Option<Vec<String>>

If specified, overrides the CMD specified in the container. If the container also has an ENTRYPOINT the values are used as entrypoint arguments. Otherwise, they are used as a command and arguments to run inside the container.

container_name: Option<String>

An optional name for the container. The container hostname will be set to this name, making it useful for inter-container communication. The name must contain only upper and lowercase alphanumeric characters and hyphens and cannot start with a hyphen.

credentials: Option<Secret>

If the specified image is hosted on a private registry other than Google Container Registry, the credentials required to pull the image must be specified here as an encrypted secret. The secret must decrypt to a JSON-encoded dictionary containing both username and password keys.

disable_image_prefetch: Option<bool>

All container images are typically downloaded before any actions are executed. This helps prevent typos in URIs or issues like lack of disk space from wasting large amounts of compute resources. If set, this flag prevents the worker from downloading the image until just before the action is executed.

disable_standard_error_capture: Option<bool>

A small portion of the container's standard error stream is typically captured and returned inside the ContainerStoppedEvent. Setting this flag disables this functionality.

enable_fuse: Option<bool>

Enable access to the FUSE device for this action. Filesystems can then be mounted into disks shared with other actions. The other actions do not need the enable_fuse flag to access the mounted filesystem. This has the effect of causing the container to be executed with CAP_SYS_ADMIN and exposes /dev/fuse to the container, so use it only for containers you trust.

entrypoint: Option<String>

If specified, overrides the ENTRYPOINT specified in the container.

environment: Option<HashMap<String, String>>

The environment to pass into the container. This environment is merged with values specified in the google.cloud.lifesciences.v2beta.Pipeline message, overwriting any duplicate values. In addition to the values passed here, a few other values are automatically injected into the environment. These cannot be hidden or overwritten. GOOGLE_PIPELINE_FAILED will be set to "1" if the pipeline failed because an action has exited with a non-zero status (and did not have the IGNORE_EXIT_STATUS flag set). This can be used to determine if additional debug or logging actions should execute. GOOGLE_LAST_EXIT_STATUS will be set to the exit status of the last non-background action that executed. This can be used by workflow engine authors to determine whether an individual action has succeeded or failed.

ignore_exit_status: Option<bool>

Normally, a non-zero exit status causes the pipeline to fail. This flag allows execution of other actions to continue instead.

image_uri: Option<String>

Required. The URI to pull the container image from. Note that all images referenced by actions in the pipeline are pulled before the first action runs. If multiple actions reference the same image, it is only pulled once, ensuring that the same image is used for all actions in a single pipeline. The image URI can be either a complete host and image specification (e.g., quay.io/biocontainers/samtools), a library and image name (e.g., google/cloud-sdk) or a bare image name ('bash') to pull from the default library. No schema is required in any of these cases. If the specified image is not public, the service account specified for the Virtual Machine must have access to pull the images from GCR, or appropriate credentials must be specified in the google.cloud.lifesciences.v2beta.Action.credentials field.

labels: Option<HashMap<String, String>>

Labels to associate with the action. This field is provided to assist workflow engine authors in identifying actions (for example, to indicate what sort of action they perform, such as localization or debugging). They are returned in the operation metadata, but are otherwise ignored.

mounts: Option<Vec<Mount>>

A list of mounts to make available to the action. In addition to the values specified here, every action has a special virtual disk mounted under /google that contains log files and other operational components. - /google/logs All logs written during the pipeline execution. - /google/logs/output The combined standard output and standard error of all actions run as part of the pipeline execution. - /google/logs/action//stdout The complete contents of each individual action's standard output. - /google/logs/action//stderr The complete contents of each individual action's standard error output.

pid_namespace: Option<String>

An optional identifier for a PID namespace to run the action inside. Multiple actions should use the same string to share a namespace. If unspecified, a separate isolated namespace is used.

port_mappings: Option<HashMap<String, i32>>

A map of containers to host port mappings for this container. If the container already specifies exposed ports, use the PUBLISH_EXPOSED_PORTS flag instead. The host port number must be less than 65536. If it is zero, an unused random port is assigned. To determine the resulting port number, consult the ContainerStartedEvent in the operation metadata.

publish_exposed_ports: Option<bool>

Exposes all ports specified by EXPOSE statements in the container. To discover the host side port numbers, consult the ACTION_STARTED event in the operation metadata.

run_in_background: Option<bool>

This flag allows an action to continue running in the background while executing subsequent actions. This is useful to provide services to other actions (or to provide debugging support tools like SSH servers).

timeout: Option<String>

The maximum amount of time to give the action to complete. If the action fails to complete before the timeout, it will be terminated and the exit status will be non-zero. The pipeline will continue or terminate based on the rules defined by the ALWAYS_RUN and IGNORE_EXIT_STATUS flags.